changed delete session so the session cant be deleted by a user who is already in a session and when a session is deleted the server checks if is has people in it if not it deletes it

+++ DatabaseManager.py
@@ -144,11 +144,14 @@ class DatabaseManager:
         cursor=self.cursor.execute("SELECT name FROM sessions WHERE name=(?)",(session,))
         result=cursor.fetchone()
         if result:
-            cursor.execute("DELETE FROM sessions WHERE name=(?)",(session,))
-            print(session)
-            print(result)
-            self.conn.commit()
-            return True
+            cursor=self.cursor.execute('SELECT username FROM active_session_users WHERE session= (?)',(session,))
+            rows=cursor.fetchall()
+            if len(rows) == 0:
+                cursor.execute("DELETE FROM sessions WHERE name=(?)",(session,))
+                print(session)
+                print(result)
+                self.conn.commit()
+                return True
         return False
